CSS Custom Filter should be a runtime enabled feature
This change removes the custom code to handle custom filters.
The feature is driven by --enable-css-shaders on Chromium but
Alexandru says it's fine to switch to a runtime enabled feature.

This also creates the first test that is only run in stable mode.
This is required as the failing test absolutely resolves the URL
in the test output.

Review URL: https://codereview.chromium.org/26085002

git-svn-id: svn://svn.chromium.org/blink/trunk@159240 bbb929c8-8fbe-4397-9dbb-9b2b20218538