Move page/animation to frame/animation
This is the "old" animation code to be replaced eventually
by the root-level animation/ directory.  That said,
AnimationController is actually owned by the Frame object
so this belongs in the frame/ directory.

I notified the Animation team that this move was coming via #blink.

BUG=305811
TBR=abarth@chromium.org

Review URL: https://codereview.chromium.org/26869007

git-svn-id: svn://svn.chromium.org/blink/trunk@159409 bbb929c8-8fbe-4397-9dbb-9b2b20218538