Fixed crash when gesture tapping node no renderer.
Image maps generate nodes in the DOM tre that don't have an attached
renderer but are returned from hit testing. This caused a crash when
attempting to display a gesture tap highlight. This fix walks up the
DOM until we find a node with a renderer.

BUG=305780

Review URL: https://codereview.chromium.org/26763007

git-svn-id: svn://svn.chromium.org/blink/trunk@159413 bbb929c8-8fbe-4397-9dbb-9b2b20218538