Provide new WebFrameClient for WebHelperPlugin.
The WebFrameClient provided by the embedder has a bunch of extra
signals that WebHelperPlugin does not need. In particular, the API
expects (or will soon expect):

 (1) that there is only 1 WebFrameImpl per client
 (2) frameDetach() messages always correspond to DOM tree nodes.

In the case of the Chromium's implementaiton of WebFrameClient, the
frameDetach() method is also entangled with lifetime management of the
associated WebFrameImpls.

This CL introduces a new HelperPluginFrameClient that only forwards
required calls from the WebHelperPlugin back up to the main client.
Having one level of indirection allows the WebHelperPlugin to
control the lifetime of its main frame, and also provides
encapsulation of the signals it sends outwards to the embedder.

BUG=306694

Review URL: https://codereview.chromium.org/27239005

git-svn-id: svn://svn.chromium.org/blink/trunk@159756 bbb929c8-8fbe-4397-9dbb-9b2b20218538