Remove introspection on GC behavior from SVG tearoff regression test
This test is really just a "test passes if it does not crash" test showing
that tearoffs can still be accessed after their elements are dead. There's
no need to use internals.numberOfLiveNodes (which is flaky due to the fact
that we run multiple tests in the same renderer) in this test.

R=pdr@chromium.org

Review URL: https://codereview.chromium.org/27741003

git-svn-id: svn://svn.chromium.org/blink/trunk@159905 bbb929c8-8fbe-4397-9dbb-9b2b20218538