Fix the position of show result category checkbox for easy access.
The option for showing category results {expected failues, flaky failures
unexpected passes & stderr} is placed at the bottom of page and is not fixed.
When the result of any category (above mentioned) is more than screen length,
we have to scroll to the end of page to change it again, which seems wrong ui.
We should have it placed along with other option (like toggle images), so that
we can change it without scrolling the page.

BUG=None

TEST=None; no behavior change;

Review URL: https://codereview.chromium.org/27488002

git-svn-id: svn://svn.chromium.org/blink/trunk@159986 bbb929c8-8fbe-4397-9dbb-9b2b20218538