Fixing TestRunner::findString() so as to fix findString.html editing testcase.
We need to call frame's stopFinding() API after the find() call. Also, false
should be passed for the |clearSelection| in parameter. This helps set focus
and selection on the found/matched range, if any.

The reason why editing/text-iterator/findString.html was failing earlier was
because selection was not being set on the matched range. Thus selection.baseOffset
and selection.extentOffset would both return zero.

Also, additional find options i.e. "StartInSelection", "AtWordStarts" and
"TreatMedialCapitalAsWordStart" are now being handled by the TestRunner
and have been added to WebFindOptions as well.

Handling these lead to a small change in behavior of LayoutTests/editing/
text-iterator/findString-restarts-at-last-position.html testcase.

The first case within this:
shouldBeTrue('testRunner.findString("first_", [])');
moves the selection to "first_".

The next case:
shouldBeTrue('testRunner.findString("first_step", [])');
was earlier returning true since we were not handling "StartInSelection" find option.
With the implementation of this option, if we wish for the next find operation
to begin from the start of the current active selection ("find_" in this case) we
should set this option explicitly.
Otherwise, if no option is passed, findNext is considered as set and we start the
next find operation from after the current active selection. Thus the aforementioned
case of finding "find_step" would return false.
To fix this, I am now passing the "StartInSelection" option to the findString() call.

Also, in the testcase:
LayoutTests/editing/text-iterator/findString.html
for the test: testFindString("webkit.org", "org", ["AtWordStarts"], [[]]);
we are currently matching the behavior for:
testFindString("webkit.org", "org", ["AtWordStarts", "TreatMedialCapitalAsWordStart"], [[7, 10], []]);
and hence I have changed the case to expect a return selection range of [7, 10].
This is the only change made in findString.html.

BUG=64773

Review URL: https://codereview.chromium.org/26654011

git-svn-id: svn://svn.chromium.org/blink/trunk@160033 bbb929c8-8fbe-4397-9dbb-9b2b20218538