Make FillLayer size type initial value settable
The initial value for FillLayer size type was SizeNone which indicates
that the size on the fill layer was unset. This means that
background-size: initial was incorrectly not setting the fill layer size
when it should have been setting it to "auto auto".

Fixes crashing test:
virtual/web-animations-css/animations/cross-fade-background-image.html

Review URL: https://codereview.chromium.org/31703002

git-svn-id: svn://svn.chromium.org/blink/trunk@160280 bbb929c8-8fbe-4397-9dbb-9b2b20218538