Remove LayoutTest animations/fill-mode-forwards.html
This test relies on receiving a specific number of animationiteration events.
However, in both the legacy and Web Animations implementations, we do not
guarantee that the correct number of these events is fired. See
crbug.com/275263. We have no immediate plans to fix this and the flakiness
dashboard shows timeouts for this test with both implementations.

The test was added to test a particular bug fix in the legacy animations
implementation (http://trac.webkit.org/changeset/117216) and has little to do
with forwards fill.

Most cases of forwards-fill testing (excluding play-state and events) is
covered by timing-model.html

This test therefore adds little value and should be removed. We can consider
reinstating it if and when the underlying bug is fixed.

BUG=275263
R=dstockwell

Review URL: https://codereview.chromium.org/29363005

git-svn-id: svn://svn.chromium.org/blink/trunk@160286 bbb929c8-8fbe-4397-9dbb-9b2b20218538